DnD: Revamped code to simplify / untangle of internal data handling: - C-ifying and renaming classes DnDURIList / DnDURIObject -> DnDTransferList / DnDTransferObject - Added testcases for DnDTransferList / DnDTransferObject + DnDPath API - Reduced memory footprint - Greatly simplified / stripped down internal data flow of Main side - More (optional) release logging for further diagnosis
Work in progress.

svn:sync-xref-src-repo-rev: r139382

+++ src/VBox/Additions/WINNT/VBoxTray/VBoxDnD.h
@@ -1,4 +1,4 @@
-/* $Id: VBoxDnD.h 85121 2020-07-08 19:33:26Z knut.osmundsen@oracle.com $ */
+/* $Id: VBoxDnD.h 85371 2020-07-17 10:02:58Z andreas.loeffler@oracle.com $ */
 /** @file
  * VBoxDnD.h - Windows-specific bits of the drag'n drop service.
  */
@@ -70,7 +70,7 @@ public:
 
     int Abort(void);
     void SetStatus(Status status);
-    int Signal(const RTCString &strFormat, const void *pvData, uint32_t cbData);
+    int Signal(const RTCString &strFormat, const void *pvData, size_t cbData);
 
 protected:
 
@@ -87,7 +87,7 @@ protected:
     RTSEMEVENT  mEventDropped;
     RTCString   mstrFormat;
     void       *mpvData;
-    uint32_t    mcbData;
+    size_t      mcbData;
 };
 
 class VBoxDnDDropSource : public IDropSource
